finish initial commands
